# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zed CV API
  description: AI-powered job matching platform for Zambia
  version: 0.1.0
  contact:
    email: convergeozambia@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.zedcv.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        phone: { type: string }
        full_name: { type: string }
        email: { type: string, format: email }
        location: { type: string }
        years_experience: { type: integer, minimum: 0 }
        skills: { type: array, items: { type: string } }
        cv_uploaded: { type: boolean }
        subscription_tier: { type: string, enum: [free, starter, professional] }
        role: { type: string, enum: [user, superadmin] }
        created_at: { type: string, format: date-time }
    UserProfileUpdate:
      type: object
      properties:
        full_name: { type: string }
        email: { type: string, format: email }
        location: { type: string }
        years_experience: { type: integer, minimum: 0 }
    UserPreferences:
      type: object
      properties:
        whatsapp_alerts: { type: boolean, default: true }
        language: { type: string, enum: [en, bem], default: en }
    UserPreferencesUpdate:
      type: object
      properties:
        whatsapp_alerts: { type: boolean }
        language: { type: string, enum: [en, bem] }
    ProfileDeleted:
      type: object
      properties:
        deleted: { type: boolean }
        user_id: { type: string, format: uuid }
    UserSkill:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        proficiency:
          type: string
          enum: [beginner, intermediate, advanced, expert]
          default: intermediate
        source:
          type: string
          enum: [cv_parse, manual, assessment]
          default: manual
    UserSkillsList:
      type: object
      properties:
        skills:
          type: array
          items: { $ref: '#/components/schemas/UserSkill' }
    UserSkillCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        proficiency:
          type: string
          enum: [beginner, intermediate, advanced, expert]
          default: intermediate
    UserSkillUpdate:
      type: object
      required: [proficiency]
      properties:
        proficiency:
          type: string
          enum: [beginner, intermediate, advanced, expert]
...
    Subscription:
      type: object
      properties:
        id: { type: string, format: uuid }
        tier: { type: string, enum: [free, starter, professional] }
        status: { type: string, enum: [active, expired, cancelled] }
        current_period_start: { type: string, format: date-time }
        current_period_end: { type: string, format: date-time }
        matches_used: { type: integer }
        matches_limit: { type: integer }
    PaymentInitiate:
      type: object
      required: [tier, payment_method]
      properties:
        tier: { type: string, enum: [starter, professional] }
        payment_method: { type: string, enum: [mtn_money, airtel_money] }
        phone: { type: string, pattern: '^\+260[0-9]{9}$' }
    PaymentInitiateResponse:
      type: object
      properties:
        transaction_token: { type: string }
        payment_url: { type: string, format: uri }
        status: { type: string, enum: [pending, redirect] }
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    patch:
      operationId: updateProfile
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserProfileUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    delete:
      operationId: deleteProfile
      summary: Hard-delete the caller's account and cascade-related rows
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileDeleted' }
  /profile/preferences:
    get:
      operationId: getPreferences
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPreferences' }
    patch:
      operationId: updatePreferences
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserPreferencesUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPreferences' }
  /profile/skills:
    get:
      operationId: listSkills
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSkillsList' }
    post:
      operationId: addSkill
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserSkillCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSkillsList' }
  /profile/skills/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    patch:
      operationId: updateSkill
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserSkillUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSkillsList' }
        '404': { description: Skill not on profile }
    delete:
      operationId: removeSkill
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSkillsList' }
        '404': { description: Skill not found }
...
  /subscription:
    get:
      operationId: getSubscription
      tags: [Subscription]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
  /subscription/pay:
    post:
      operationId: initiatePayment
      tags: [Subscription]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInitiate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentInitiateResponse' }
...